Create auth actions, effects, and reducer
