+++ src/app/effects/auth.effects.ts
@@ -9,30 +9,28 @@ import 'rxjs/add/operator/mergeMap';
 import 'rxjs/add/operator/switchMap';
 import { Observable } from 'rxjs/Observable';
 
-import * as AppActions from '../actions/app.actions';
+import * as AuthActions from '../actions/auth.actions';
 import { OrganizationService } from '../services/organization.service';
-import { SnackBarService } from '../services/snack-bar.service';
 import { UserService } from '../services/user.service';
 
 @Injectable()
-export class AppEffects {
+export class AuthEffects {
 
   @Effect()
   loadData$: Observable<Action> = this.actions
-    .ofType(AppActions.LOAD_DATA)
-    .map((action: AppActions.LoadData) => action.payload)
+    .ofType(AuthActions.LOAD_DATA)
+    .map((action: AuthActions.LoadData) => action.payload)
     .switchMap(fbUser => this.userService.getById(fbUser.uid)
       .switchMap(res => {
         const user = Object.assign({}, res, { email: fbUser.email, id: fbUser.uid });
         return this.organizationService.getById(user.organizationId)
           .map(organization => {
             const org = Object.assign({}, organization, { id: user.organizationId });
-            return new AppActions.LoadDataSuccess({ user: user, organization: org} )
+            return new AuthActions.LoadDataSuccess({ user: user, organization: org })
           })
       }));
 
   constructor(private actions: Actions,
-              private snackBarService: SnackBarService,
               private userService: UserService,
               private organizationService: OrganizationService) {}
 }
